Multiple Changes
Added functions to manage paging
Added more memory functions
Added support for vector(primitive)
Refactored PCI

+++ src/kernel/std/Bitmap.cpp
@@ -4,35 +4,35 @@
 
 namespace std
 {
-    Bitmap::Bitmap(uint8_t* _bitmap, uint32_t bitcount, bool defaultValue) : bitmap(_bitmap)
+    Bitmap::Bitmap(uint8_t* _bitmap, size_t bitcount, bool defaultValue) : bitmap(_bitmap)
     {
         bitmap_size = DivRoundUp(bitcount, 8);
         bitmap_count = bitcount;
-        for(uint32_t i = 0; i < bitmap_size; i++)
+        for(size_t i = 0; i < bitmap_size; i++)
         {
             bitmap[i] = 0xFF * (uint8_t)defaultValue;
         }
     }
 
-    bool Bitmap::get(uint32_t bitIndex)
+    bool Bitmap::get(size_t bitIndex)
     {
-        uint32_t byteIndex = bitIndex / 8;
-        uint32_t byteOffset = bitIndex % 8;
+        size_t byteIndex = bitIndex / 8;
+        size_t byteOffset = bitIndex % 8;
 
         return (bitmap[byteIndex] >> byteOffset) & 0x1;
     }
-    void Bitmap::set(uint32_t bitIndex, bool value)
+    void Bitmap::set(size_t bitIndex, bool value)
     {
-        uint32_t byteIndex = bitIndex / 8;
-        uint32_t byteOffset = bitIndex % 8;
+        size_t byteIndex = bitIndex / 8;
+        size_t byteOffset = bitIndex % 8;
 
         if(value) bitmap[byteIndex] |= (1 << byteOffset);
         else bitmap[byteIndex] &= ~(1 << byteOffset);
     }
 
-    uint32_t Bitmap::find_false()
+    size_t Bitmap::find_false()
     {
-        for(uint32_t i = 0; i < bitmap_size; i++)
+        for(size_t i = 0; i < bitmap_size; i++)
         {
             uint8_t value = bitmap[i];
 
@@ -57,9 +57,9 @@ namespace std
         return Bitmap::npos;
     }
 
-    uint32_t Bitmap::find_false_bits(uint32_t n)
+    size_t Bitmap::find_false_bits(size_t n)
     {
-        for(uint32_t i = 0; i < bitmap_size; i++)
+        for(size_t i = 0; i < bitmap_size; i++)
         {
             uint8_t value = bitmap[i];
 
@@ -73,9 +73,9 @@ namespace std
                     // the >> will givew a even number
                     if((value >> offset) % 2 == 0)
                     {
-                        uint32_t loc = i * 8 + offset;
+                        size_t loc = i * 8 + offset;
 
-                        for(uint32_t j = 0; j < n; j++)
+                        for(size_t j = 0; j < n; j++)
                         {
                             // not a false byte
                             // skip
@@ -91,16 +91,17 @@ namespace std
             }
 
             nextByte:
+            ;
         }
 
         return Bitmap::npos;
     }
 
-    void Bitmap::setBits(uint32_t first, uint32_t size, bool value)
+    void Bitmap::setBits(size_t first, size_t size, bool value)
     {
         if(size < 8)
         {
-            for(uint32_t i = 0; i < size; i++)
+            for(size_t i = 0; i < size; i++)
             {
                 set(first + i, value);
             }
@@ -109,10 +110,10 @@ namespace std
         }
 
         uint8_t byteValue = 0xFF * (uint8_t)value;
-        uint32_t byteCount = size / 8;
+        size_t byteCount = size / 8;
 
-        uint32_t firstByte = first / 8;
-        uint32_t lastByte = (first + size) / 8;
+        size_t firstByte = first / 8;
+        size_t lastByte = (first + size) / 8;
 
         uint8_t beginOffset = first % 8;
         uint8_t lastOffset = 8 - ((first + size) % 8);
@@ -131,7 +132,7 @@ namespace std
 
         //bitmap[firstByte] |= byteValue << beginOffset;
 
-        for(uint32_t i = firstByte + 1; i < lastByte; i++)
+        for(size_t i = firstByte + 1; i < lastByte; i++)
         {
             bitmap[i] = byteValue;
         }
@@ -139,7 +140,7 @@ namespace std
         bitmap[lastByte] = (bitmap[lastByte] & ~(0xFF >> lastOffset)) | (byteValue >> lastOffset);
     }
 
-    uint32_t Bitmap::size()
+    size_t Bitmap::size()
     {
         return bitmap_count;
     }
