+++ src/kernel/std/Bitmap.hpp
@@ -7,28 +7,28 @@ namespace std
     class Bitmap
     {
         private:
-            uint32_t bitmap_size = 0;   // size in bytes
-            uint32_t bitmap_count = 0;
+            size_t bitmap_size = 0;   // size in bytes
+            size_t bitmap_count = 0;
             uint8_t* bitmap = nullptr;
         
         public:
-            static const uint32_t npos = 0;
+            static const size_t npos = 0;
 
             Bitmap() = default;
-            Bitmap(uint8_t* _bitmap, uint32_t bitcount, bool defaultValue = true);
+            Bitmap(uint8_t* _bitmap, size_t bitcount, bool defaultValue = true);
 
-            bool get(uint32_t bitIndex);
-            void set(uint32_t bitIndex, bool value);
+            bool get(size_t bitIndex);
+            void set(size_t bitIndex, bool value);
 
-            void setBits(uint32_t first, uint32_t size, bool value);
+            void setBits(size_t first, size_t size, bool value);
 
             // finds bit whose value is false
-            uint32_t find_false();
+            size_t find_false();
 
             // finds n consecutive bits which are false
-            uint32_t find_false_bits(uint32_t n);
+            size_t find_false_bits(uint32_t n);
 
-            uint32_t size();
+            size_t size();
 
             ~Bitmap() = default;
     };
